refactor: KoreaEntryViewController 파일을 Controller 폴더로 이동 및 리팩토링
